Module faucet::faucet

The Espresso Faucet

Structs

FaucetOptions

FaucetQueue ğŸ”’

A shared, asynchronous queue of requests.

FaucetQueueIndex ğŸ”’

FaucetState ğŸ”’

HealthCheck

Enums

FaucetStatus

Functions

break_up_records ğŸ”’

Break records into smaller pieces to create at least state.num_records total.

check_service_available ğŸ”’

healthcheck ğŸ”’

Return a JSON expression with status 200 indicating the server is up and running. The JSON expression is simply, {"status": Status} where Status is one of

init_web_server

faucet_key_pair - If provided, will be added to the faucet keystore.

main ğŸ”’

maintain_enough_records ğŸ”’

Worker task to maintain at least state.num_records in the faucet keystore.

request_fee_assets ğŸ”’

spendable_records ğŸ”’

worker ğŸ”’