Function faucet::faucet::maintain_enough_records

async fn maintain_enough_records(state: FaucetState, wakeup: Receiver<()>)

Worker task to maintain at least state.num_records in the faucet keystore.

When signalled on wakeup, this thread will break large records into small records of size state.grant_size, until there are at least state.num_records distinct records in the keystore.

The record breakup is only triggered when the number of available records is less than half of the desired number of records, and in that case we always replenish all the way to the desired number of records if possible. This prevents us from generating a record transaction every time we do a transfer, and ensures that whenever we do break up records, we break up many at a time, so we can take advantage of the parallelism of having multiple record breakup transactions in flight at the same time.